Upped version to 0.8.2.
Changed calculation of hash size for jellyfish.  Genome sizes are now handled with longs rather than ints.

+++ src/main/java/uk/ac/tgac/rampart/tool/process/analyse/asm/selector/AssemblySelector.java
@@ -14,6 +14,6 @@ import java.io.File;
 public interface AssemblySelector {
 
     File selectAssembly(AssemblyStatsTable table,
-                        int estimatedGenomeSize,
+                        long estimatedGenomeSize,
                         double estimatedGcPercentage);
 }
